requirements.txt
llama-parse
torch
torch-vision
transformers
python-dotenv